+++ packages/app/src/cli/models/extensions/function-specifications/delivery_customization.ts
@@ -4,6 +4,7 @@ const spec = createFunctionSpecification({
   identifier: 'delivery_customization',
   externalIdentifier: 'delivery_customization',
   externalName: 'Function - Delivery customization',
+  supportedFlavors: [{name: 'Javascript', value: 'vanilla-js'}],
   gated: false,
   templatePath: (lang) => `checkout/${lang}/delivery-customization/default`,
 })
