+++ packages/app/src/cli/services/deploy/upload.ts
@@ -37,7 +37,11 @@ export async function uploadThemeExtensions(
         registrationId: themeId,
       }
       const mutation = api.graphql.ExtensionUpdateDraftMutation
-      const result: api.graphql.ExtensionUpdateSchema = await api.partners.request(mutation, token, themeExtensionInput)
+      const result: api.graphql.ExtensionUpdateSchema = await api.partners.partnersRequest(
+        mutation,
+        token,
+        themeExtensionInput,
+      )
       if (result.extensionUpdateDraft?.userErrors?.length > 0) {
         const errors = result.extensionUpdateDraft.userErrors.map((error) => error.message).join(', ')
         throw new error.Abort(errors)
@@ -95,7 +99,11 @@ export async function uploadUIExtensionsBundle(
   }
 
   const mutation = api.graphql.CreateDeployment
-  const result: api.graphql.CreateDeploymentSchema = await api.partners.request(mutation, options.token, variables)
+  const result: api.graphql.CreateDeploymentSchema = await api.partners.partnersRequest(
+    mutation,
+    options.token,
+    variables,
+  )
 
   if (result.deploymentCreate?.userErrors?.length > 0) {
     const errors = result.deploymentCreate.userErrors.map((error) => error.message).join(', ')
@@ -125,7 +133,11 @@ export async function getUIExtensionUploadURL(apiKey: string, deploymentUUID: st
     bundleFormat: 1,
   }
 
-  const result: api.graphql.GenerateSignedUploadUrlSchema = await api.partners.request(mutation, token, variables)
+  const result: api.graphql.GenerateSignedUploadUrlSchema = await api.partners.partnersRequest(
+    mutation,
+    token,
+    variables,
+  )
   if (result.deploymentGenerateSignedUploadUrl?.userErrors?.length > 0) {
     const errors = result.deploymentGenerateSignedUploadUrl.userErrors.map((error) => error.message).join(', ')
     throw new error.Abort(errors)
