+++ packages/app/src/cli/services/dev/create-extension.ts
@@ -29,7 +29,7 @@ export async function createExtension(
     config: JSON.stringify({}),
     context: null,
   }
-  const result: api.graphql.ExtensionCreateSchema = await api.partners.request(query, token, variables)
+  const result: api.graphql.ExtensionCreateSchema = await api.partners.partnersRequest(query, token, variables)
 
   if (result.extensionCreate.userErrors?.length > 0) {
     const errors = result.extensionCreate.userErrors.map((error) => error.message).join(', ')
