+++ packages/app/src/cli/services/dev/fetch.ts
@@ -48,7 +48,7 @@ export async function fetchAppExtensionRegistrations({
   apiKey: string
 }): Promise<api.graphql.AllAppExtensionRegistrationsQuerySchema> {
   const query = api.graphql.AllAppExtensionRegistrationsQuery
-  const result: api.graphql.AllAppExtensionRegistrationsQuerySchema = await api.partners.request(query, token, {
+  const result: api.graphql.AllAppExtensionRegistrationsQuerySchema = await api.partners.partnersRequest(query, token, {
     apiKey,
   })
   return result
@@ -62,7 +62,7 @@ export async function fetchAppExtensionRegistrations({
  */
 export async function fetchOrganizations(token: string) {
   const query = api.graphql.AllOrganizationsQuery
-  const result: api.graphql.AllOrganizationsQuerySchema = await api.partners.request(query, token)
+  const result: api.graphql.AllOrganizationsQuerySchema = await api.partners.partnersRequest(query, token)
   const organizations = result.organizations.nodes
   if (organizations.length === 0) throw NoOrgError()
   return organizations
@@ -78,7 +78,7 @@ export async function fetchOrgAndApps(orgId: string, token: string, title?: stri
   const query = api.graphql.FindOrganizationQuery
   const params: {id: string; title?: string} = {id: orgId}
   if (title) params.title = title
-  const result: api.graphql.FindOrganizationQuerySchema = await api.partners.request(query, token, params)
+  const result: api.graphql.FindOrganizationQuerySchema = await api.partners.partnersRequest(query, token, params)
   const org = result.organizations.nodes[0]
   if (!org) throw NoOrgError(orgId)
   const parsedOrg = {id: org.id, businessName: org.businessName, appsNext: org.appsNext}
@@ -86,19 +86,23 @@ export async function fetchOrgAndApps(orgId: string, token: string, title?: stri
 }
 
 export async function fetchAppFromApiKey(apiKey: string, token: string): Promise<OrganizationApp | undefined> {
-  const res: api.graphql.FindAppQuerySchema = await api.partners.request(api.graphql.FindAppQuery, token, {apiKey})
+  const res: api.graphql.FindAppQuerySchema = await api.partners.partnersRequest(api.graphql.FindAppQuery, token, {
+    apiKey,
+  })
   return res.app
 }
 
 export async function fetchOrgFromId(id: string, token: string): Promise<Organization | undefined> {
   const query = api.graphql.FindOrganizationBasicQuery
-  const res: api.graphql.FindOrganizationBasicQuerySchema = await api.partners.request(query, token, {id})
+  const res: api.graphql.FindOrganizationBasicQuerySchema = await api.partners.partnersRequest(query, token, {id})
   return res.organizations.nodes[0]
 }
 
 export async function fetchAllDevStores(orgId: string, token: string): Promise<OrganizationStore[]> {
   const query = api.graphql.AllDevStoresByOrganizationQuery
-  const result: api.graphql.AllDevStoresByOrganizationSchema = await api.partners.request(query, token, {id: orgId})
+  const result: api.graphql.AllDevStoresByOrganizationSchema = await api.partners.partnersRequest(query, token, {
+    id: orgId,
+  })
   return result.organizations.nodes[0]!.stores.nodes
 }
 
@@ -119,7 +123,7 @@ export async function fetchStoreByDomain(
   shopDomain: string,
 ): Promise<FetchStoreByDomainOutput | undefined> {
   const query = api.graphql.FindStoreByDomainQuery
-  const result: api.graphql.FindStoreByDomainSchema = await api.partners.request(query, token, {
+  const result: api.graphql.FindStoreByDomainSchema = await api.partners.partnersRequest(query, token, {
     id: orgId,
     shopDomain,
   })
