+++ packages/app/src/cli/services/generate.test.ts
@@ -130,7 +130,7 @@ async function mockSuccessfulCommandExecution(identifier: string, existingExtens
   })
 
   vi.mocked(loadApp).mockResolvedValue(app)
-  vi.mocked(api.partners.request).mockResolvedValueOnce({extensionSpecifications: testRemoteSpecifications})
+  vi.mocked(api.partners.partnersRequest).mockResolvedValueOnce({extensionSpecifications: testRemoteSpecifications})
   vi.mocked(ensureGenerateEnvironment).mockResolvedValue('api-key')
   vi.mocked(generateExtensionPrompt).mockResolvedValue({name: 'name', extensionType: identifier})
   vi.mocked(generateExtensionService).mockResolvedValue(path.join(appRoot, 'extensions', 'name'))
