--- packages/cli-kit/src/api/common.ts
@@ -1,98 +1,0 @@
-import {firstPartyDev} from '../environment/local.js'
-import constants from '../constants.js'
-import {stringifyMessage, content, token as outputToken, token, debug} from '../output.js'
-import {Abort, ExtendableError} from '../error.js'
-import {ClientError, RequestDocument, Variables} from 'graphql-request'
-import {randomUUID} from 'crypto'
-
-export class RequestClientError extends ExtendableError {
-  statusCode: number
-  public constructor(message: string, statusCode: number) {
-    super(message)
-    this.statusCode = statusCode
-  }
-}
-
-export async function buildHeaders(token?: string): Promise<{[key: string]: string}> {
-  const userAgent = `Shopify CLI; v=${await constants.versions.cliKit()}`
-
-  const headers: {[header: string]: string} = {
-    'User-Agent': userAgent,
-    // 'Sec-CH-UA': secCHUA, This header requires the Git sha.
-    'Sec-CH-UA-PLATFORM': process.platform,
-    'X-Request-Id': randomUUID(),
-    'Content-Type': 'application/json',
-    ...(firstPartyDev() && {'X-Shopify-Cli-Employee': '1'}),
-  }
-  if (token) {
-    // eslint-disable-next-line dot-notation
-    headers['authorization'] = `Bearer ${token}`
-    headers['X-Shopify-Access-Token'] = `Bearer ${token}`
-  }
-
-  return headers
-}
-
-/**
- * Removes the sensitive data from the headers and outputs them as a string.
- * @param headers - HTTP headers.
- * @returns A sanitized version of the headers as a string.
- */
-export function sanitizedHeadersOutput(headers: {[key: string]: string}): string {
-  const sanitized: {[key: string]: string} = {}
-  const keywords = ['token', 'authorization']
-  Object.keys(headers).forEach((header) => {
-    if (keywords.find((keyword) => header.toLocaleLowerCase().includes(keyword)) === undefined) {
-      sanitized[header] = headers[header]!
-    }
-  })
-  return Object.keys(sanitized)
-    .map((header) => {
-      return ` - ${header}: ${sanitized[header]}`
-    })
-    .join('\n')
-}
-
-export function debugLogRequest<T>(
-  api: string,
-  query: RequestDocument,
-  variables?: Variables,
-  headers: {[key: string]: string} = {},
-) {
-  debug(content`
-Sending ${token.json(api)} GraphQL request:
-${token.raw(query.toString())}
-
-With variables:
-${variables ? JSON.stringify(variables, null, 2) : ''}
-
-And headers:
-${sanitizedHeadersOutput(headers)}
-`)
-}
-
-export async function handlingErrors<T>(api: string, action: () => Promise<T>): Promise<T> {
-  try {
-    return await action()
-  } catch (error) {
-    if (error instanceof ClientError) {
-      const errorMessage = stringifyMessage(content`
-  The ${token.raw(
-    api,
-  )} GraphQL API responded unsuccessfully with the HTTP status ${`${error.response.status}`} and errors:
-
-  ${outputToken.json(error.response.errors)}
-      `)
-      let mappedError: Error
-      if (error.response.status < 500) {
-        mappedError = new RequestClientError(errorMessage, error.response.status)
-      } else {
-        mappedError = new Abort(errorMessage)
-      }
-      mappedError.stack = error.stack
-      throw mappedError
-    } else {
-      throw error
-    }
-  }
-}
